+++ doc/ChangeLog.md
@@ -1,5 +1,8 @@
 # Changelog for GPSFTP5 scripts
 
+## 2019-12-16	sjm	Removed StatusDB.pm
+- Removed StatusDB.pm again. Use individual horly status files and status.0 in exclusive mode.
+
 ## 2019-12-15	sjm	Use threads in gpspickup
 - Make gpspickup multi-threaded to parallelize the unpacking process
 - Increased the wait time for multi-file incoming file sets.